Move "findParentDomain" out of store/options into utils
